requirements.txt
opencv-python
Pillow
numpy
rembg
